# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Bayyina CRM API
  description: |
    REST API for Bayyina CRM — an education center management system.

    ## Authentication
    All protected endpoints require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    Tokens are obtained via `POST /api/auth/login` or `POST /api/auth/verify-otp`.
    Tokens expire after **30 days**.

    ## Roles
    | Role      | Description                                      |
    |-----------|--------------------------------------------------|
    | `admin`   | Full access to all resources                     |
    | `teacher` | Read access to students; manage groups/payments  |
    | `student` | Read-only access to own data                     |

    ## Error Format
    All errors follow this structure:
    ```json
    { "code": "errorCode", "message": "Human readable message" }
    ```

  version: 1.0.0
  contact:
    name: Mardonbek Khamidov
    email: mardonbek@bayyina.uz
servers:
  - url: http://localhost:4000/api
    description: Local development
  - url: https://api.bayyina.uz/api
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued on login or OTP verification. Valid for 30 days.

  # ─────────────────────────────────────────────
  # Reusable schemas
  # ─────────────────────────────────────────────
  schemas:
...
    RejectionReason:
      type: object
      description: A configurable reason for rejecting a lead.
      properties:
        _id:
          type: string
          example: "664f1a2b3c4d5e6f7a8b9c12"
        title:
          type: string
          description: Short title of the rejection reason
          example: "Narx juda qimmat"
        description:
          type: string
          nullable: true
          description: Additional detail
          example: "Lead said the price is too high compared to competitors"
        createdBy:
          oneOf:
            - type: string
              nullable: true
            - type: object
              properties:
                _id: { type: string, example: "664f1a2b3c4d5e6f7a8b9c0f" }
                firstName: { type: string, example: "Sarvinoz" }
                lastName: { type: string, example: "Yusupova" }
        createdAt:
          type: string
          format: date-time
          example: "2026-04-01T08:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-04-01T08:00:00.000Z"

    # ── Notification ─────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "invalidCredentials"
        message:
          type: string
          example: "Telefon raqam yoki parol noto'g'ri"

    # ── Record ──────────────────────────────────
...
  parameters:
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Page number (1-based)
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
      description: Number of results per page (default 20, max 100)
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            tokenMissing:
              summary: No token provided
              value: { code: "tokenMissing", message: "Token mavjud emas" }
            tokenExpired:
              summary: Token has expired
              value: { code: "tokenExpired", message: "Token muddati tugagan" }
            invalidToken:
              summary: Token is malformed
              value: { code: "invalidToken", message: "Yaroqsiz token kiritildi" }
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "forbidden"
            message: "Ruxsat berilmadi"
...
    BadRequest:
      description: Validation error or bad input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded (120 req / 3 min)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "tooManyRequests"
            message: "Juda ko'p so'rovlar. Iltimos, 3 daqiqadan keyin urinib ko'ring"
...
paths:
...
  /rejection-reasons:
    get:
      tags: [Rejection Reasons]
      summary: List rejection reasons
      description: Returns a paginated list of rejection reasons. Supports free-text search (`q`). **Admin only.**
      operationId: getRejectionReasons
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
        - name: q
          in: query
          description: Search by title or description
          schema:
            type: string
            example: "narx"
      responses:
        "200":
          description: Rejection reasons returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  rejectionReasons:
                    type: array
                    items:
                      $ref: "#/components/schemas/RejectionReason"
                  total: { type: integer, example: 6 }
                  page: { type: integer, example: 1 }
                  limit: { type: integer, example: 20 }
                  totalPages: { type: integer, example: 1 }
                  hasNextPage: { type: boolean, example: false }
                  hasPrevPage: { type: boolean, example: false }
                  code: { type: string, example: rejectionReasonsFound }
                  message: { type: string, example: "Rad etish sabablari ro'yxati olindi" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    post:
      tags: [Rejection Reasons]
      summary: Create a rejection reason
      description: Creates a new rejection reason. **Admin only.**
      operationId: createRejectionReason
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  example: "Narx juda qimmat"
                description:
                  type: string
                  example: "Lead said the price is too high compared to other centers"
      responses:
        "201":
          description: Rejection reason created
          content:
            application/json:
              schema:
                type: object
                properties:
                  rejectionReason:
                    $ref: "#/components/schemas/RejectionReason"
                  code: { type: string, example: rejectionReasonCreated }
                  message: { type: string, example: "Rad etish sababi muvaffaqiyatli yaratildi" }
        "400":
          description: Missing title
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example: { code: "missingField", message: "Rad etish sababi kiritilishi shart" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /rejection-reasons/{id}:
    get:
      tags: [Rejection Reasons]
      summary: Get a rejection reason by ID
      operationId: getRejectionReason
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c12"
      responses:
        "200":
          description: Rejection reason found
          content:
            application/json:
              schema:
                type: object
                properties:
                  rejectionReason:
                    $ref: "#/components/schemas/RejectionReason"
                  code: { type: string, example: rejectionReasonsFound }
                  message: { type: string, example: "Rad etish sabablari ro'yxati olindi" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Rejection reason not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example: { code: "rejectionReasonNotFound", message: "Rad etish sababi topilmadi" }
        "429":
          $ref: "#/components/responses/TooManyRequests"

    put:
      tags: [Rejection Reasons]
      summary: Update a rejection reason
      operationId: updateRejectionReason
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c12"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Boshqa markazni tanladi"
                description:
                  type: string
                  example: "Went to a competitor"
      responses:
        "200":
          description: Rejection reason updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  rejectionReason:
                    $ref: "#/components/schemas/RejectionReason"
                  code: { type: string, example: rejectionReasonUpdated }
                  message: { type: string, example: "Rad etish sababi muvaffaqiyatli yangilandi" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Rejection reason not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example: { code: "rejectionReasonNotFound", message: "Rad etish sababi topilmadi" }
        "429":
          $ref: "#/components/responses/TooManyRequests"

    delete:
      tags: [Rejection Reasons]
      summary: Delete a rejection reason
      operationId: deleteRejectionReason
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c12"
      responses:
        "200":
          description: Rejection reason deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string, example: rejectionReasonDeleted }
                  message: { type: string, example: "Rad etish sababi muvaffaqiyatli o'chirildi" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Rejection reason not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example: { code: "rejectionReasonNotFound", message: "Rad etish sababi topilmadi" }
        "429":
          $ref: "#/components/responses/TooManyRequests"