# src/models/RejectionReason.js
const mongoose = require("mongoose");

const RejectionReason = new mongoose.Schema(
  {
    title: { type: String, required: true, trim: true },
    description: { type: String, trim: true },
    createdBy: {
      type: mongoose.Schema.Types.ObjectId,
      ref: "User",
    },
  },
  { timestamps: true },
);

module.exports = mongoose.model("RejectionReason", RejectionReason);